Refactor catalog class hierarchy
I consolidated all catalog subclasses into one and used a factory method
and state subclasses to define a catalog. Under the new design, a user
can create a catalog using a single URI or a list of URIs.

Also, the new catalog class will watch the URI(s) for updates and,
optionally, cache the results for a specified time period. This removes
the need for helper and callback methods in the server. A client will
also benefit from these changes.