New parameter templates for local shell/env
The idea is to have some shell/env values pulled from the client
environment before sending to remote cat server.

In the case that the same cat is used locally, the new templates
will evaluate as before, with the difference that permissions are
not required (since the user already has arbitrary access by
definition).