Initial attempt
adds 'persist()' method to sources, which uses the container lookup to
find the appropriate thing to call for that source. Saves to one
particular format for each container type, only dataframe implemented so
far here.

Also includes GenericDataFrame, which can load anything given a set of
files and a function which can make an open file into a data-frame piece.