Improve UserParameter __repr__.
In #255 we wanted a UserParameter repr that was more useful but not too
verbose. We settled on making `__str__` a complete eval-able
representation but leaving `__repr__` as the not-so-descriptive `object`
defaut.

I propose a different compromise: unify `__str__` and `__repr__` but use
a succinct non-eval-able repr, `<UserParameter 'some_name_here'>` where,
as in the standard library `<>` is used to indicate the the repr is not
eval-able but merely a summary.

Reasoning:

* An eval-able representation is nice, but our current implementation of
eval-able `__str__` is backward: normally it's the `__repr__` that is
(where possible) eval-able. If we agree that the eval-able string is too
verbose, maybe it would be better to settle on short but still useful
repr.
* Python _supports_ different `__str__` and `__repr__` but to my taste
it's nice when they can be consistent. When presented with a
not-so-useful `__repr__`, it doesn't always occur to me to try printing
it to see if the `__str__` representation is better.

Closes #465