MNT: clean up remaining leakage of CatalogEntry to public API
The `items` method was yielding directly from `cat._entries`, this
converts them to DataSource objects before returning them to the user.
This in turn required converting some internal calls to use
`self._entries` directly.

The `cat.filter` method takes in a function that operates directly on
CatalogEntry rather than a DataSource.  I don't see an easy way to
convert this API away from CatalogEntry so left a warning in the
doctoring.