Checked for `skipfooter` csv kwarg in CSVSource._get_schema
Ensuring that setting `skipfooter` as a csv argument in CSVSource does not
lead to a crash when `nrows` is also set in `CSVSource._get_schema`. The
solution in this commit unsets `skipfooter` before trying to read the schema,
but adds `skipfooter` lines to `nrows`, and then ignores the final `skipfooter`
rows in the returned DataFrame, to ensure that any footers don't mess up the
schema inference.