Fixed schema inference problem for CSVSource with skipfooter.
The code from the previous change would not infer the right dtypes for files
that had less than 10+skipfooter rows. For such small files, the footers would
be included when `pandas.read_csv` inferred dtypes. The code in this commit
fixes that problem by re-reading the file and re-inferring dtypes if it
detects that `skipfooter` is not None, and the total number of rows in the
DataFrame read by CSVSource._get_schema is less than `10+skipfooter`.