Changed how skipfooter CSV files are stored and accessed for testing.
This change was necessary to prevent conflicts with tests that rely on the
content of source.tests.test_csv.data_filenames conforming to the original
content, or that don't tolerate extra CSV files being added to
intake/source/tests.

After the change, all tests in `intake/source/tests/test_csv.py` pass, but many
other tests fail. Need to rerun tests in a different branch to confirm whether
this is caused by the new code, or whether it is a problem in the original
master branch.