Add handling of environment variables of type str
ast.literal_eval() does not parse strings and either raises a ValueError or a SyntaxError.
Thus, when a environment variable of the form `INTAKE_*="any-string"` is present `val` would be undefined and `self[k2] = val` would raise an UnboundLocalError.

This will catch both possible errors and depending on a new config setting "raise_on_error" will either parse the environment variable as a string while giving a warning (default behaviour) or raise.