Pinned h5py in pip env due to HDF error in test
The new test I added was raising an HDF error on CI.
This seems to only be for linux. The error is noted on this
issue: https://github.com/Unidata/netcdf4-python/issues/1438

I added h5py as a dependency and pinned it to < 3.15, which seems to
fix the error.